refactor: move SortedSegIntersections to intersect tool
